Americanization now applied per buffer.
If you are working on several documents in different Vim buffers, it may
be that you require American spellings in one, and British spellings in
another.

Previously, sourcing the americanize.vim file once would mean
that the "corrections" were available in all buffers. Now, you have to
explicitly source the americanize.vim file in each buffer where you want
access to the abbreviations.