Customizations for ctags/taglist.
Add .ctags file to the list of 'dotfiles' that should be symlinked.

The .ctags file includes extensions to add support for markdown and PML
filetypes.

Also, taglist plugin needs to be made aware of these customizations, so
vimrc includes tlist_xxx_settings variables, which tell it how to deal
with the extenstions to ctags.

More info on customizing ctags/taglist:

    http://ctags.sourceforge.net/EXTENDING.html
    http://vim-taglist.sourceforge.net/extend.html