Remove 'local' settings for ignore/smart-cases.
I was confused about why smartcase and ignorecase appeared to be set,
even though I didn't specify them in my vimrc file. It turned out that
the Vimoutliner was setting them, but using `setlocal`. If it was
possible to set these values locally to a buffer, then this wouldn't
have caused any trouble, but both of these are global settings. So as
soon as I opened an .otl file, ignorecase and smartcase were set for all
buffers.

Glad I got to the bottom of that one.